requirements.txt
django ==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
Pillow==7.1.2
python-decouple==3.3
django-hitcount==1.3.2 
django-tinymce == 3.0.2